Add dw2-aranges.exp
This test checks that GDB is able to load DWARF information when
.debug_aranges has a section address size that is set to 0.

This test was originally written by Jan Kratochvil to test commit
af0aa51e5b0 from 2017, titled "DWARF-5: .debug_names index consumer".

This test was originally written using a static .S file and has
been present in the Fedora tree for a long time.

If dwarf2/aranges.c is modified to turn off the address_size check,
GDB will crash with SIGFPE when loading the executable with address
size set to zero.

I modified the DWARF assembler to make it possible to set the address
size to zero in a .debug_aranges section and used the DWARF assembler
to produce the assembly file.

Co-Authored-By: Jan Kratochvil <jan.kratochvil@redhat.com>
Approved-by: Kevin Buettner <kevinb@redhat.com>