gdb/testsuite: fix test pattern after switch to -lbl matching
After commit:

  commit 938a5652011fa47ace19874eb5847add9b1e11fa
  Date:   Fri Oct 25 06:14:03 2024 +0200

      [gdb/testsuite] Fix some test-cases for check-read1 (-lbl)

I notice that gdb.base/sect-cmd.exp would sometimes fail.  The problem
is that by switching to line by line matching we now need to ensure
that the gdb_test_multiple patterns match up to the end of the line,
but don't actually include the trailing \r\n (yeah, our line by line
matching is weird).  We need to be especially careful anywhere '.*' is
used as this can potentially match content on a subsequent line.

I have replaced '.*' with '\[^\r\n\]*(?=\r\n)', matching everything up
to the end of the line, but not the end of line itself, and I've made
use of '(?=\r\n)' in a couple of other places to ensure we match up to
the end of the line, but don't match the line terminator itself.